Старт переделки SettingsView
