--- HSMonitor/Views/Settings/SettingsSideHostWindow.xaml
@@ -1,32 +1,0 @@
-<Window x:Class="HSMonitor.Views.SettingsSideHostWindow"
-        xmlns="http://schemas.microsoft.com/winfx/2006/xaml/presentation"
-        xmlns:x="http://schemas.microsoft.com/winfx/2006/xaml"
-        xmlns:materialDesign="http://materialdesigninxaml.net/winfx/xaml/themes"
-        WindowStyle="None"
-        AllowsTransparency="True"
-        Background="Transparent"
-        ShowInTaskbar="False"
-        ResizeMode="NoResize"
-        WindowStartupLocation="Manual"
-        SnapsToDevicePixels="True">
-    <Border CornerRadius="5"
-            Background="{DynamicResource PrimaryHueLightBrush}"
-            SnapsToDevicePixels="True">
-
-        <materialDesign:DialogHost
-            x:Name="RootDialogHost"
-            materialDesign:TransitionAssist.DisableTransitions="True"
-            Style="{DynamicResource MaterialDesignEmbeddedDialogHost}"
-            DialogBackground="Transparent"
-            Background="Transparent"
-            DialogTheme="Light"
-            ApplyBlurBackground="True"
-            HorizontalAlignment="Stretch"
-            HorizontalContentAlignment="Stretch"
-            VerticalAlignment="Stretch"
-            VerticalContentAlignment="Stretch"
-            DialogMargin="0">
-            <Grid />
-        </materialDesign:DialogHost>
-    </Border>
-</Window>
